Backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
tensorflow==2.13.0
pillow==10.0.0